Patch ImagePullSecrets to CoreDNS and Kube-proxy ServiceAccount
When upgrading Kubernetes from version 1.29 to 1.30, the ServiceAccount
configuration for CoreDNS and Kube-proxy pods is reset. This reset
causes the previously configured imagePullSecrets to be removed,
resulting in ImagePullBackOff errors on compute nodes.

To address this issue, we are patching the imagePullSecrets
into the CoreDNS and Kube-proxy serviceaccount after the
control plane upgrade

Test plan:
PASS: build-pkgs && build-image.
PASS: Perform k8s upgrade from 1.29 to 1.30, and then to 1.31 on
AIO-SX and verify it upgraded successfully.
PASS: Perform k8s upgrade from 1.29 to 1.30, and then to 1.31 on
AIO-DX and verify it upgraded successfully.
PASS: Perform k8s upgrade from 1.29 to 1.30, and then to 1.31
on a standard system and verify coredns and kube-proxy are running
on compute nodes.
PASS: Perform multiple k8s upgrade on AIO-SX and verify it
upgraded successfully.

Closes-bug: 2106037

Change-Id: Idf7a6451951b28617336e98f5385d6674deef448
Signed-off-by: Boovan Rajendran <boovan.rajendran@windriver.com>