Configure ksoftirqd threads priority
This commit aims to allow the configuration and tuning of the ksoftirqd
priority. This is done by adding a new parameter to the platform.conf
file, which is used by affine-platform script to set the priority.

TEST PLAN:
(Test executed in AIO machines with, and without, lowlatency
configuration)
PASS: Bootstrap system.
PASS: Success on adding and removing the ksoftirqd parameter from the
      platform.conf file.
PASS: Success on setting the ksoftirqd priority.

Depends-On: https://review.opendev.org/c/starlingx/utilities/+/947501

Story: 2011401
Task: 52106

Change-Id: I12917777e4a96acaeca56f7843986bbf7642cc74
Signed-off-by: Eduardo Alberti <eduardo.alberti@windriver.com>