Add include section to IPsec swanctl.conf file
This commit adds an implementation to include IPsec configuration files
into swanctl.conf file. A section containing all includes is added to
the end of swanctl.conf file, after the connections specification.

This is introduced as a requirement from the pod-to-pod IPsec
encryption implementation where the /etc/swanctl/conf.d/*.conf must be
included in swanctl.conf file, as this file is intended to have all
specifications defined for connections loaded on StrongSwan.

Test Paln:
PASS: Build puppet-manifests, ipsec-auth, sysinv packages.
PASS: Build, install and deploy a DX system with a worker and a storage
      nodes containing these changes. Observe that swanctl.conf file is
      generated with the 'include conf.d/*.conf' spec at the EOF.
PASS: Install ipsec-policy-operator app, create an IPsecPolicy and wait
      until IPsec Policy Agent generates k8s-node.conf files for each
      kubernetes node. Reload IPsec connections and observe that conns
      under k8s-nodes.conf file are reloaded.

Story: 2011127
Task: 52088

Depends-On: https://review.opendev.org/c/starlingx/app-ipsec-policy-operator/+/948485
Change-Id: I0fecbf6adc248fb045064b6db7c2c1465af399a6
Signed-off-by: Manoel Benedito Neto <Manoel.BeneditoNeto@windriver.com>