Avoid removing registry.central from /etc/hosts until unlock
The registry.central entry is added during bootstrap to enable name
resolution before the dnsmasq service is running. However, this entry
was being removed by the puppet manifest because it wasn’t explicitly
declared, and "resources { 'host': purge => true }" was applied
unconditionally.
As a result, after the deployment manager yaml is applied, but before
the system is unlocked, the entry is purged from /etc/hosts. If any
issue occurs during this window, it may cause communication failures
with registry.central, potentially impacting operations such as image
retrieval or configuration.
This change ensures that "purge => true" is only triggered after the
/etc/platform/.unlock_ready flag file is present, preventing
premature removal of registry.central from /etc/hosts during the
critical bootstrap to unlock window.

Test Plan
=========

[PASS] Fresh install of a DC system with a subcloud. After deployment is
applied and before the unlock, verify that the registry.central entry
remains in /etc/hosts. It should only be removed after the unlock.

[PASS] Fresh install of an SX system. After deployment is applied and
before the unlock, verify that the registry.central entry does not exist
in /etc/hosts, and that no issues occur before or after the unlock.

Closes-Bug: 2110563

Change-Id: Iccb8a13fe0efcc14a476eb2ef26baac43c49e942
Signed-off-by: Ferdinando Terada <Ferdinando.GodoyTerada@windriver.com>