Delete pods with OnDelete during root CA update
During a Kubernetes root-CA update, pods must restart to pick up the
new cluster CA certs. Previously, only a rollout restart was triggered.
Rollout restarts do not apply to DaemonSets or StatefulSets with the
OnDelete strategy, causing an error if one of those is present.

With this change, pods are deleted directly when the strategy is
OnDelete, rather than invoking a rollout restart.

Test Plan:
PASS: Run build-pkgs -c -p puppet-manifests
PASS: Run build-image
PASS: Deploy the built image, have a DaemonSet with the OnDelete
      updateStrategy, run the kube-rootca-update-pods.erb script
      directly and verify that all pods are restarted with no timeout.
PASS: Run kube-rootca-update procedure in the scenario above and watch
      the OnDelete pods restart.
PASS: Run kube-rootca-update strategy in a DC central with a out-of-sync
      subcloud and watch the OnDelete pods restart.

Closes-Bug: 2111751

Change-Id: I2c2c0fd45d093cfe76d0b182b233911efa2bc0fc
Signed-off-by: Rodrigo Tavares <Rodrigo.DosSantosTavares@windriver.com>