Fix condition to restart dccertmon service
Currently, the `onlyif` parameter uses a boolean expression instead of
a command string, causing Puppet to fail. As a result, the Kubernetes
Root CA update procedure cannot complete.

This change wraps the condition in a shell test command, ensuring the
dccertmon service restarts only on the systemcontroller.

Test Plan:
PASS: Run build-pkgs -c -p puppet-manifests
PASS: Run build-image
PASS: Run kube-rootca-update procedure in an AIO-SX and wait for it to
      complete without the dccertmon restart step.
PASS: Run kube-rootca-update procedure in an DC system controller and
      wait for it to complete with the dccertmon restart step.

Change-Id: If0db206a415f7329196441c5cfdfddfdd6c267fa
Signed-off-by: Rodrigo Tavares <Rodrigo.DosSantosTavares@windriver.com>