Update hugepage mount location on Debian
In addition to the information shared at [1], in order to support
OVS-DPDK as the vswitch type on Debian following the puppet vswitch
upgrade, it is necessary to adjust the hugepage directory
configuration. When attempting to use /mnt directly as the hugepage
path, a "permission denied" error occurs during DPDK initialization.
To solve this issue, the same approach used for the puppet filesystem
at [2] and [3] is being applied, since /mnt is a symlink to
/var/rootdirs/mnt.

[1] https://review.opendev.org/c/starlingx/integ/+/951436
[2] https://opendev.org/starlingx/stx-puppet/commit/d43047ba3
[3] https://opendev.org/starlingx/stx-puppet/commit/d92d79361

TEST PLAN:

PASS - build-pkgs -c -p puppet-manifests
PASS - build-pkgs --all
PASS - build-image
PASS - AIO-SX fresh install with OVS as vswitch type

Depends-On: https://review.opendev.org/c/starlingx/stx-puppet/+/951437

Story: 2011263
Task: 52381

Change-Id: I3e4f5ce73d0a6316bff0b48a8afb0bd85bec8837
Signed-off-by: Murillo Arantes <murillo.arantes@windriver.com>