Change Ceph script for active_cluster_updates class
The migration procedure from AIO-SX to AIO-DX, with Ceph storage
backend configured, is failing because active_cluster_updates class
is using the ceph-init-wrapper script to start/stop Ceph services.

The ceph-init-wrapper script checks for the existence of the
/var/run/.ceph_started flag before proceeding with the required
actions. However, it's created only after mtcClient is started.

Since this puppet manifests is executed before, it is returning
a false positive, return code 0, and does not start the services,
which leads to blockling further cephfs operations.

Using the /etc/init.d/ceph script directly, it is possible to perform
the same operations, without going through the wrapper script.

Test Plan:
  PASS: AIO-SX -> AIO-DX Migration: No storage backend.
  PASS: AIO-SX -> AIO-DX Migration: Ceph storage backend.
  PASS: AIO-SX -> AIO-DX Migration: Rook-Ceph storage backend.

Closes-bug: 2115275

Change-Id: I952d60b91959a77eb96981b4b91a929352a700e1
Signed-off-by: Hediberto C Silva <hediberto.cavalcantedasilva@windriver.com>