Fix cyclic dependency between networking and rate limit classes
The cyclic dependency error occurred during rate_limit configuration
or when converting the network setup from single stack to dual stack.
The issue was caused by the inclusion of the platform::networking
anchor inside the platform::network::interfaces::rate_limit class,
which created a loop in the dependency graph and caused Puppet to
fail:

Error: Found 1 dependency cycle:
(Anchor[platform::networking] => Class[
Platform::Network::Interfaces::Rate_limit] => Class[
Platform::Network::Interfaces::Rate_limit] => Stage[pre] => Stage[
main] => Class[Platform::Anchors] => Anchor[platform::networking])

The fix consists of including the rate_limit class only within
the platform::network::apply class, just like other network classes such
as addresses, interfaces, and routes.

Additionally, the checks for the haproxy and vim-webserver services
were revised to address issues encountered during dual stack
configuration. The vim-webserver check was removed, as the service
was disabled due to being unused and unnecessarily consuming memory.
For haproxy, the status check timeout was increased to prevent
failures caused by premature timeout while the service was still
being updated, which previously led to configuration errors during
dual stack setup.

Test Plan
=========

[PASS] On a single stack IPv6 SX system, configure rate-limit for both
TX and RX on a platform interface and verify that the configuration is
correctly applied to ip6tables. Then, convert the system to dual stack
IPv6/IPv4 and check that the configuration is properly reflected in the
cluster. Also verify that the rate-limit configuration is now applied
to both iptables and ip6tables.

[PASS] On a dual stack IPv4/IPv6 DX system, remove the secondary
address pool associated with the cluster and verify that the cluster
is reconfigured correctly to single stack.

[PASS] On a single stack IPv4 DX system, convert the network to dual
stack IPv4/IPv6, then configure TX and RX rate-limiting on the
platform interface. Verify that the dual stack configuration is
applied by creating dual stack pods and confirming connectivity to
both IPv4 and IPv6 endpoints. Also verify that the rate-limit
configuration is properly applied to both iptables and ip6tables.
Then, perform a lock and unlock of one of the controllers and verify
that both the rate-limit and dual stack configurations persist.

[PASS] Perform bootstrap on a dual stack IPv6/IPv4 SX system. Verify
the dual stack configuration in the cluster by creating pods and
ensuring connectivity to service endpoints over both IPv4 and IPv6.

[PASS] On a single stack IPv6 SX system, convert to dual stack
IPv6/IPv4 and verify that the dual stack configuration is reflected in
the cluster. Then, configure rate-limit on a platform interface and
check that it is applied to both iptables and ip6tables. Lock and
unlock the controller and ensure that the rate-limit and dual stack
configurations persist.

[PASS] On a single stack IPv4 DX system, convert to dual stack
IPv4/IPv6 with rate-limiting already configured on a platform
interface. Verify that the configuration is applied correctly to the
cluster and that rate-limiting is configured in both iptables and
ip6tables.

[PASS] Perform bootstrap on a single stack IPv4 DX system. Verify that
the cluster is configured with IPv4 stack only.

Story: 2011350
Task: 52447

Change-Id: Idcf24d65580e4c136e86ac05d8bd2a7475a78d37
Signed-off-by: Ferdinando Terada <Ferdinando.GodoyTerada@windriver.com>