Kube-apiserver port change upgrade rollback
During the upgrade to stx 11, the port for kube-apiserver will be
changing from 6443 to 16443. This change needs to be reverted in
case the upgrade script that applies it was already executed
before the abort/rollback.

This review introduces code to detect the rollback and not include
the new haproxy configuration, use the correct port when
reconfiguring the k8s control plane and clean the flags used.

Test plan:
PASS: Upgrade activation - AIO-SX and DX (stx10->11).
PASS: Rollback after upgrade activation - AIO-SX and DX (stx11->10).

Story: 2011399
Task: 52473

Change-Id: Iecd3b741858abea496c3bd40c8c2eab135c9c116
Signed-off-by: Marcelo de Castro Loebens <Marcelo.DeCastroLoebens@windriver.com>