Set k8s_update.log file permissions to 640
Currently, /var/log/puppet/XXX_runtime/k8s_update.log
have permission 644. To comply with the CIS benchmark
requirements, the permissions should be set to 640.

This change updates the permissions of
/var/log/puppet/XXX_runtime/k8s_update.log to 640.

The chmod is applied immediately after the FileHandler
is instantiated to ensure the permission change takes
effect as soon as the file is created.

Test Plan:
PASS: build-pkgs and build-image
PASS: Verify that permission of k8s_update.log is
      set to 640
      Ran below commands to generate the k8s_update.log
      - system service-parameter-modify kubernetes \
      kube_apiserver default-not-ready-toleration-seconds=33
      - system service-parameter-apply kubernetes

Story: 2011241
Task: 52507

Change-Id: I03bae36cf0c888cc3accca11b0fa86738dc25c16
Signed-off-by: Rahul Roshan Kachchap <rahulroshan.kachchap@windriver.com>