Add missing include section in swanctl.conf
This commit adds a class responsible to write at the end of swanctl
configuration files an include section to import .conf files. This
class aims to address a missing include section to swanctl.conf during
system upgrades from stx.10 to stx.11.

Test Plan:
PASS: Build an image with the changes for puppet-manifests.
PASS: In a DX system, execute the class manually and observe that an
      include section is added at the end of /etc/swanctl/
      stongswan_active.conf and /etc/swanctl/stongswan_standby.conf.
      Observe that /etc/swanctl/swanctl.conf is updated accordingly.
PASS: In the same DX system, after the first run, execute the puppet
      class again manually and observe that the included section is
      not repeated in the file and remain added in all swanctl
      configuration files.

Story: 2011127
Task: 52545

Change-Id: I29d2c05f32cf242761041ddae8f19dfe11bce0bb
Signed-off-by: Manoel Benedito Neto <Manoel.BeneditoNeto@windriver.com>