Fix strongswan config during upgrades
This commit fixes the swanctl.conf reconfigure per the addition or
removal of an include section at the end of /etc/swanctl/swanctl.conf
file during upgrades to stx.11. Except for the controller nodes, all
other nodes need to have the include section addressed only on the
/etc/swanctl/swanctl.conf file.

Test Plan:
PASS: Build puppet-manifests package.
PASS: In a DX+ system, execute the include_files class manually and
      observe that an include section is added at the end of
      /etc/swanctl/swanctl.conf file for compute nodes.
PASS: In a DX+ system, execute the exclude_files class manually and
      observe that the include section is removed from the end of
      /etc/swanctl/swanctl.conf file for compute nodes.
PASS: In a DX+ system, manually execute the upgrade script #27 and
      observe that the include section is added at the end of
      /etc/swanctl/swanctl.conf file for compute nodes for activate
      action. All other nodes have their swanctl.conf file updated
      accordingly.
PASS: In a DX+ system, manually execute the upgrade script #27 and
      observe that the include section is removed from the end of
      /etc/swanctl/swanctl.conf file for compute nodes for activate
      rollback action. All other nodes have their swanctl.conf file
      updated accordingly.

Story: 2011127
Task: 52644
Closes-Bug: 2119688

Change-Id: I4d473c85376a6487b56bc32098a45909d4229bed
Signed-off-by: Manoel Benedito Neto <Manoel.BeneditoNeto@windriver.com>