K8s external requests with client certs (admin)
With the introduction of haproxy in ssl bridge mode between requests
from external clients and kube-apiserver, client certificates are
not able to reach kube-apiserver anymore (ssl termination for the
client request is done in haproxy, and a new ssl connection is made
between haproxy and kube-apiserver).

This change allows external clients that provide client certificates
with credentials for the user kubernetes-admin to connect with
kube-apiserver. The external client can only access kube-apiserver
with client certificate if the credentials described in the subject
are: "O = kubeadm:cluster-admins, CN = kubernetes-admin"

This is made by having a separate backend where haproxy itself
authenticates with admin credentials to kube-apiserver, while
verifying the external client certificate in the frontend before
redirecting to it.

Test plan:
PASS: Deploy AIO-SX.
      Tested external requests to kube-apiserver with client
      certificates and kubernetes-admin credentials.
      Tested external requests to kube-apiserver with client
      certificates and non kubernetes-admin credentials (negative
      test).
      Tested external requests to kube-apiserver with token.
      Run manually the script kube-cert-rotation.sh, verified
      that the bundle file with the client certificate for
      haproxy was updated, and haproxy restarted and is working
      as expected.

PASS: Deploy DC SystemController + SX subcloud
      Tested external requests to kube-apiserver with client
      certificates and kubernetes-admin credentials.
      Tested external requests to kube-apiserver with client
      certificates and non kubernetes-admin credentials (negative
      test).
      In SystemController, verified that the bundle file with the
      client certificate for haproxy is in place in controller-1,
      and haproxy is working as expected after a swact.
      Performed procedure to update the kubernetes root ca.

Story: 2011399
Task: 52646

Change-Id: I257cf13b3df768a740ca584acc8f50d84466f8c0
Signed-off-by: Marcelo de Castro Loebens <Marcelo.DeCastroLoebens@windriver.com>