Increase retries for apply_network_config
During boot, the Puppet code runs the script apply_network_config.py
to configure the platform interfaces.
This step modifies interface settings and toggles interfaces down/up.
At the same time, the sysinv code may also need access to the interfaces
for example, to perform LLDP operations or read Interface PCI attributes.
For correct operation, these actions use a mutex to ensure that interfaces
are not modified by another service while an operation is in progress.

In some scenarios, apply_network_config.py could not acquire the mutex even
after 5 retries with 5 second intervals.
As a result, Puppet failed, and the node configuration did not complete
successfully.

Analysis showed that during sysinv's inventory_audit, the mutex is
frequently locked for 4–7 seconds, particularly during the ports
inventory in pci_get_net_attrs.

To address this, the number of retries was increased, which resolved
the issue.
An analysis to improve the sysinv code and reduce the sysinv agent's
operation time is in progress.

Tests Done:

Fresh install on a standard lab setup reproducing the error

Partial-Bug: 2120572
Change-Id: I5f031b668de89f6aa487f51da74ac8ba5803f05c
Signed-off-by: Fabiano Correa Mercer <fabiano.correamercer@windriver.com>