Fix race condition during dc-vault mount
When controller-0 is deployed and the dc-vault is configured, if
dc-vault was already mounted (by systemd/fstab automount) while only
controller-0 is being deployed, Puppet attempts to mount dc-vault
again due to the automount flag. This causes Puppet to return an error
indicating that the filesystem is already mounted.

To resolve this, the automount option for dc-vault was disabled in the
Puppet configuration.

Test plan:

STD and DX (DC)

- PASS: run bootstrap and perform unlock on controler-0
- PASS: after unlock controller-0, before installing the hosts,
        perform 3 reboots
- PASS: complete the installation and run a DOR test

Closes-bug: 2121539

Change-Id: I3def039f8c4eb2175ecfb0ddfc4412e7da19ffc1
Signed-off-by: Gabriel Przybysz Gonçalves Júnior <gabriel.przybyszgoncalvesjunior@windriver.com>