Add proxy frontend for the admin network to HAProxy
The admin network needs to be treated just like the mgmt
network when it comes to subclouds deployed with it.

This commit adds a HAProxy frontend for the admin network
and port 6443 in TCP mode so that internal requests will reach
kube-apiserver directly.

Test Plan:
PASS: build-pkgs & build-image.
PASS: On an AIO-SX subcloud:
      - Bootstrap the system with ipv4
      - Bootstrap the system with ipv6
      - On a running subcloud add the admin network
      and do dcmanager update on the system controller
      to change the subcloud management network to be
      the newly defined admin network
      - Verify that admin ip address are added to the
      kube-apiserver certificate.
      - Send curl requests from the system controller
      to subcloud as https://[SUBCLOUD-ADMIN-IP]:6443
      and verifies it reaches kube-apiserver directly.

Story: 2011399
Task: 52780

Signed-off-by: Rei Oliveira <Reinildes.JoseMateusOliveira@windriver.com>
Change-Id: I04d7ee38ef2c8f6a4d6d1b0c92d72df1246afa77