Add admin ip addresses to kube-apiserver certificate
The admin network needs to be treated just like the mgmt
network when it comes to subclouds deployed with it.

As such we need to have the admin ip address properly added
to the kube-apiserver certificate so that communication can
be done over the admin network.

In addition, this review fixes 2 code mistakes in conditionals
to check if c0 and c1 addresses where empty. This was
discussed with networking SMEs and they agreed.

Test Plan:
PASS: build-pkgs & build-image.
PASS: On an AIO-SX subcloud:
      - Bootstrap the system with ipv4
      - Bootstrap the system with ipv6
      - On a running subcloud add the admin network
      and do dcmanager update on the system controller
      to change the subcloud management network to be
      the newly defined admin network
      - Verify that admin ip address are added to the
      kube-apiserver certificate.
      - Send curl requests from the system controller
      to subcloud as https://[SUBCLOUD-ADMIN-IP]:6443
      and verifies it reaches kube-apiserver directly.

Story: 2011399
Task: 52781

Signed-off-by: Rei Oliveira <Reinildes.JoseMateusOliveira@windriver.com>
Change-Id: I5cc3d22945c4d310e6a197778c735010de35ebfd