Remove cloud init file after enrollment configuration
The removal of the /etc/network/interfaces.d/50-cloud-init
file was introduced after the enrollment operation is completed, since
keeping this file in the system would cause configuration
inconsistencies. For example, if there is a new OAM reconfiguration,
the address from 50-cloud-init would still persist after a lock/unlock
cycle, which would consequently lead to configuration inconsistencies.

Test Plan
=========

After enrollment, check that the 50-cloud-init file no longer exists
under /etc/network/interfaces.d. This verification must be performed
as part of all test cases.

[PASS] Execute subcloud enrollment process against a DX subcloud by
changing the bootstrap values IP to addresses from different subnets
than those configured in the subcloud and switching the bootstrap
interface from one VLAN X to another VLAN y. Verify that the default
route is updated and that the IP addresses on the VLAN are correct
after the unlock.

[PASS] Execute subcloud enrollment process against a DX subcloud by
changing the bootstrap values IP to addresses from different subnets
than those configured in the subcloud and switching the bootstrap
interface from VLAN to physical. Verify that the default route is
updated and that the IP addresses on the VLAN are correct after the
unlock.

[PASS] Execute subcloud enrollment process against a DX subcloud by
changing the bootstrap values IP to addresses from different subnets
than those configured in the subcloud and switching the bootstrap
interface from VLAN to physical. Verify that the default route is
updated and that the IP addresses on the VLAN are correct after the
unlock.

[PASS] Execute subcloud enrollment process against a DX subcloud by
changing the bootstrap values IP to addresses from different subnets
than those configured in the subcloud. Verify that the default route
is updated and that IP addresses on the VLAN are correct before and
after the unlock.

[PASS] Execute subcloud enrollment process against a DX subcloud by
changing the bootstrap values IP to a different IP from the one
configured in the subcloud but within the same subnet (varying the
floating and node addresses). Verify that the default route, as well
as the floating and IP addresses on the VLAN, are correct before and
after the unlock.

[PASS] Execute subcloud enrollment process against a DX subcloud
by changing the bootstrap values IP to match the configured subcloud
IP and VLAN. Verify that the default route, as well as the IP addresses
on the VLAN, are correct before and after the unlock.

[PASS] Execute subcloud enrollment process against a SX subcloud by
changing the bootstrap values IP to addresses from different subnets
than those configured in the subcloud and switching the bootstrap
interface from VLAN to physical. Verify that the default route is
updated and that the IP addresses on the VLAN are correct after the
unlock.

[PASS] Execute subcloud enrollment process against a SX subcloud by
changing the bootstrap values IP to a different IP from the one
configured in the subcloud but within the same subnet (varying the
floating and node addresses). Verify that the default route, as well
as IP addresses on the VLAN, are correct before and after the unlock.

Closes-Bug: 2121954

Change-Id: I167a60fd15af3cbdcc46e20af58a23760f94e180
Signed-off-by: Ferdinando Terada <Ferdinando.GodoyTerada@windriver.com>