Fix deny rules for k8s admin client certificate
For external access to kube-apiserver (via OAM network), only tokens
are supported. The only exception is users with client certificates
with the k8s admin credentials. This certificate need to have the
following fields in the subject:
"O = kubeadm:cluster-admins, CN = kubernetes-admin"

Requests with client certificates that doesn't have both of these
fields should be denied by HAproxy. Previous logic was not right,
if the certificate had one of the fields, it wouldn't be denied, so
it required a fix.

Test plan:
PASS: Bootstrap AIO-SX.
PASS: Make a request to kube-apiserver via OAM using a valid token.
      Observe that the request succeeds.
PASS: Make a request to kube-apiserver via OAM using a client
      certificate w/ subject:
      "O = kubeadm:cluster-admins, CN = Any"
      Observe that request is denied.
PASS: Make a request to kube-apiserver via OAM using a client
      certificate w/ subject:
      "O = Any, CN = CN=kubernetes-admin"
      Observe that request is denied.
PASS: Make a request to kube-apiserver via OAM using a client
      certificate w/ subject:
      "O = kubeadm:cluster-admins, CN = CN=kubernetes-admin"
      Observe that the request succeeds.

Story: 2011399
Task: 52805

Change-Id: I7654ec706fd8f19bf17d57f75663e710739248eb
Signed-off-by: Marcelo de Castro Loebens <Marcelo.DeCastroLoebens@windriver.com>