ts2phc instance stops working after reboot
This change makes sure that the SDP pins are configured before the
ts2phc service starts running.

The instances for ts2phc were not working properly after reboot, but
when puppet was applied via 'system ptp-instance-apply', the service
worked correctly again.

We figured this was the case because the 'system ptp-instance-apply'
calls a runtime script inside puppet, that creates dependencies for
nic_clock and ptpinstance, but the path that the boot puppet
initialization takes does not have this dependency.
Our solution was to insert this dependency inside the ptpinstance
class, so that the SDP pins get configured before the ts2phc service.

Test Plan:
PASS: Fresh install and bootstrap
PASS: system application-apply
PASS: Verify no alarms

Closes-Bug: 2125931

Change-Id: I2bc14435b58f763c30274e9540f447f25748519b
Signed-off-by: Bruno Veiga <Bruno.VeigadeAlmeida@windriver.com>