Remove kickstart's DHCP lease file from controller nodes
In controller nodes, the pxeboot interface is statically configured.
However, during PXE installation, DHCP is used. Since mtcClient first
searches for the lease file and then, if on a controller node, checks
the static configuration, failure to remove the lease file after the
first unlock causes it to always consult the lease file — which may or
may not contain an address matching the static configuration. This
change ensures the lease file is removed at the same time as other
leftover files from the kickstart script.

This change also adds
- miliseconds to the log format message
- the network info will have the log prefix info

Test Plan
=========
Tests done for IPv4 and IPv6 STANDARD installations
[PASS] install with mgmt as ethernet interface
[PASS] install with mgmt as bonding interface
[PASS] install with pxeboot as ethernet interface and mgmt as vlan
[PASS] install with pxeboot as bonding interface and mgmt as vlan

Closes-Bug: 2127081
Depends-On: https://review.opendev.org/c/starlingx/config/+/963401

Change-Id: If55480580064ed9f80d849ef5554f3a036e386c8
Signed-off-by: Andre Kantek <andrefernandozanella.kantek@windriver.com>