Use GlobalNetworkSet in the firewall
This change introduces the use of GlobalNetworkSets to store DC routes
added to the firewall. This approach prevents the creation of large
GlobalNetworkPolicies in the system controller, which previously
occurred due to the high number of managed subclouds. The oversized
policies were causing the kube-API to reject the configuration
because of its size in bytes.

A GlobalNetworkSet is a Calico resource that can hold all subnets to
be used in the source or destination fields of a policy rule. These
sets are linked using a selector, which matches its value against
the labels defined in the GlobalNetworkSet.

All firewall bash scripts received a better handling of input
arguments, as well testing if kube-API is available (if not marks to
be applied later by sysinv) before using kubectl.

Test Plan
=========
[PASS] Install a non-DC AIO-DX dual-stack system, add routes to the
        management interface, and verify that no GlobalNetworkSet
        is created.
[PASS] Install a system controller AIO-DX dual-stack system, add
        routes to the management interface, and verify that a
        GlobalNetworkSet is created.
[PASS] Install a subcloud AIO-SX dual-stack system, add routes to the
        management interface, and verify that a GlobalNetworkSet is
        created.
[PASS] Install a subcloud AIO-SX dual-stack system, add routes to the
        admin interface, and verify that a GlobalNetworkSet is
        created.
[PASS] Validate firewall functionality between the subcloud and system
        controller.

Story: 2011311
Task: 53250

Depends-On: https://review.opendev.org/c/starlingx/config/+/967532
Change-Id: Ic4072b5e819d58fc234b610e19a7a95ae76dd95b
Signed-off-by: Andre Kantek <andrefernandozanella.kantek@windriver.com>