Ignore hosts update during mgmt reconfig process
The /etc/hosts update is part of "class platform::config::pre",
which means it runs whenever any other runtime puppet is executed.
When an AIO-SX is locked and the management address pool changes,
some puppet actions run and /etc/hosts is updated with the new
management IP. However, this IP is not yet active; it is applied
only after the host is unlocked.

Updating /etc/hosts before the unlock does not affect PODs or most
critical modules, but it impacts the fault management subsystem.
If an operation triggers an alarm set or clear, the FM code uses
C++ getaddrinfo, which resolves the controller address from
/etc/hosts. Because the new IP is not active, the FM connection
fails by timeout (5 minutes). This occurs, for example, when a
platform interface is deleted.

To avoid this failure, /etc/hosts will not be updated until the
host is unlocked after a management address pool update.

Tests Done:
- AIO-SX fresh install
- AIO-DX fresh install
- DC with one AIO-SX subcloud
- AIO-SX management reconfiguration;
    node rebooted and achieved the enabled-available state
- AIO-SX subcloud management reconfiguration
    node rebooted and achieved the enabled-available state

Closes-Bug: 2133126

Change-Id: If1c656497944402ace76e3a701fcc17125c1722a
Signed-off-by: Fabiano Correa Mercer <fabiano.correamercer@windriver.com>