Restore sriov enable device file
The removal of sriov.enable-device.erb caused a regression because
devices.pp still depends on this template. Its absence triggered the
error "Error while evaluating a Function Call, Could not find template
'platform/sriov.enable-device.erb'". The file must be kept because,
although it is no longer used in network.pp, it is still required by
the class platform::devices::fpga::fec::pf.

The change that removed the file is referenced here:
https://review.opendev.org/c/starlingx/stx-puppet/+/967721

Closes-Bug: 2133229

Change-Id: Iecff23d92760d60c2717a1a3bebfa8076830cd7f
Signed-off-by: Ferdinando Terada <Ferdinando.GodoyTerada@windriver.com>