LVM CSI support
This commit is part of a group of modifications whose objective is to
support the new StarlingX application "LVM CSI APP".

During the system configuration, to use the new app, the user can create
new volume groups and assign one of "thin" or "thick" provisioning
modes. When the "thin" provisioning mode is selected, a thin pool
logical volume will be part of the volume group.

Also, the CGTS-VG can now be modified to support a thin pool, and this
pool can be resized (only extended) as the user wants.

Considering this, this change adds a new puppet function to create and
delete a thin pool, adds a class to configure a new volume group for
this purpose and modifies the cgts-vg to include a thin pool and manage
it at runtime.

Test Plan:
PASS: host locked and unlocked successfully
PASS: new volume group, with thick provisioning mode, is successfully
      created and deleted
PASS: new volume group, with thin provisioning mode, is successfully
      created and deleted
PASS: cgts-vg is successfully changed and a new thin pool is
      successfully created and deleted
PASS: thin pool, from cgts-vg, is successfully extended

Story: 2011699
Task: 53995

Author: Eduardo Alberti <eduardo.alberti@windriver.com>
Co-author: Hediberto C Silva <Hediberto.CavalcantedaSilva@windriver.com>

Change-Id: I5f319caba822d3070adef99d59470ad31a183261
Signed-off-by: Eduardo Alberti <eduardo.alberti@windriver.com>